+++ src/main/java/com/jinjjaseoul/auth/model/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.jinjjaseoul.auth.model;
 
-import com.jinjjaseoul.domain.user.model.User;
-import com.jinjjaseoul.domain.user.model.UserRepository;
+import com.jinjjaseoul.domain.user.model.entity.User;
+import com.jinjjaseoul.domain.user.model.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
