+++ src/main/java/com/jinjjaseoul/domain/user/model/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.jinjjaseoul.domain.user.model;
+package com.jinjjaseoul.domain.user.model.repository;
 
+import com.jinjjaseoul.domain.user.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
